moved pagination from producitontype to its own pagnationtype
