+++ frontend/app/features/archive/types/paginationTypes.ts
@@ -1,0 +1,9 @@
+export interface PaginationRequest {
+  cursor?: number;
+  limit?: number;
+}
+
+export interface PaginationResponse {
+  next_cursor?: number;
+  has_more: boolean;
+}
